+++ odoo/local-src/sale_cancel_remaining/wizard/cancel_remaining_wizard.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 # Copyright 2018 Okia SPRL
+# Copyright 2018 Jacques-Etienne Baudoux (BCIM) <je@bcim.be>
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
 from odoo import api, models, _
@@ -17,36 +18,30 @@ class CancelRemainingWizard(models.TransientModel):
             raise UserError(_('No sale order line ID found'))
         line = self.env['sale.order.line'].browse(active_id)
 
-        procurements = line.procurement_ids
-        moves = procurements.mapped('move_ids')
-        moves_qty = sum([move.product_uom_qty for move in moves])
-
-        if line.product_qty_remains_to_deliver != moves_qty:
-            raise UserError(
-                _('The remaining quantity on the sale order should be the '
-                  'same than the quantity on the delivery picking'))
-
-        moves_state = set([move.state for move in moves])
-        right_state = {'draft', 'waiting', 'confirmed', 'assigned'}
-        if moves_state - right_state:
-            raise UserError(_('You cannot cancel a quantity that is part '
-                              'of a started picking'))
-
         internal_pickings = line.order_id.picking_ids.filtered(
-            lambda picking: picking.picking_type_code == 'internal')
-
-        internal_stock_move = self.env['stock.move'].search([
-            ('picking_id', 'in', internal_pickings.ids),
-            ('product_id', '=', line.product_id.id),
-            ('state', 'in', ['draft', 'waiting', 'confirmed', 'assigned']),
-            ('picking_id.printed', '=', False)
-        ])
-
-        if not internal_stock_move:
+            lambda picking: picking.picking_type_code == 'internal'
+            and picking.state not in ('cancel', 'done'))
+        if not internal_pickings:
+            raise UserError(_('No picking can be canceled'))
+        if True in internal_pickings.mapped('printed'):
             raise UserError(_('You cannot cancel a quantity that is part '
                               'of a started picking'))
 
-        procurements.cancel()
+        cancel_moves = line.procurement_ids.mapped('move_ids').filtered(
+                lambda m: m.state not in ('done', 'cancel'))
+
+        def _descend_moves(lvl):
+            next_lvl = lvl.mapped('move_orig_ids')
+            if next_lvl:
+                lvl |= _descend_moves(next_lvl)
+            return lvl
+
+        if cancel_moves:
+            cancel_moves = _descend_moves(cancel_moves)
+            cancel_moves.with_context(cancel_procurement=True).action_cancel()
+            # This will mark the procurement as done
+            cancel_moves.mapped('procurement_id').check()
+
         line.write({
             'product_qty_canceled': line.product_qty_remains_to_deliver
         })